==> Notebooks/01_Data_understanding.ipynb <==
## **A Data-Driven Market Intelligence System for Stock Risk and Behavior Analysis at the Nairobi Securities Exchange**

## Problem Statement

Despite the Nairobi Securities Exchange (NSE) seeing a surge in equity turnover-rising 18% to Ksh 56 billion in 2025 ([SE Half-Year Results, 2025](https://www.nse.co.ke/wp-content/uploads/NSE-Plc-Unaudited-Group-results-for-the-6-months-ended-30-June-2025.pdf)), retail participation remains hampered by a critical information gap. According to the Institute of Certified Investment and Financial Analysts reveals that 77% of Kenyan retail investors rely on "personal research" and social intuition because they lack accessible analytical tools ([ICIFA Annual Report, 2024](https://icifa.co.ke/static/resources/others/annual-report-2024465e3dbed42d.pdf)).

While the market added over Ksh 1 Trillion in capitalization since 2023, most investors suffer from "herding behavior," where decisions are made by following the crowd rather than technical data -[USIU-Africa Research, 2025](https://erepo.usiu.ac.ke/xmlui/bitstream/handle/11732/8460/MASILA%20BRIAN%20SALU%20MBA%202024.pdf?sequence=1&isAllowed=y). This project bridges this gap by converting raw daily prices into behavioral risk clusters, moving investors from intuition to evidence-based decision-making.

The dataset used for this analysis is publicly available from Mendeley: [Nairobi Securities Exchange (NSE) Kenya](https://data.mendeley.com/research-data/?query=Nairobi%20Securities%20Exchange%20(NSE)%20Kenya%20-%20All%20Stocks%20Prices)

## Objectives

**Main Objective**
- How can a data-driven stock market intelligence system be developed for the Nairobi Securities Exchange to support informed and risk-aware investment decisions?

**Specific Objectives**
1) `Feature Engineering:` How can financial metrics such as Rolling Volatility, Daily Returns, and Maximum Drawdowns be derived from historical stock price data to effectively quantify stock behavior?

2) `Behavioral Segmentation:` How can unsupervised machine learning techniques (e.g., K-Means and DBSCAN) be applied to group NSE-listed stocks into risk-based clusters such as stable, high-volatility, and speculative stocks?

4) `Sector Risk Analysis:` What systemic risks and stability patterns can be identified across different market sectors within the Nairobi Securities Exchange?

5) `Interactive Deployment:` How can the derived insights be interactively deployed through a Streamlit dashboard to enable users to select stocks, assess their risk profiles, and compare them against sector-level benchmarks?

**Solution**

This project aims to provide an interactive dashboard where investors can explore NSE stocks, assess risk clusters, and compare sector performance. Users can make data-driven decisions, spot stable or high-risk stocks, and identify where to invest confidently.

## Libraries & Importations

import pandas as pd
import numpy as np

df_2021 = pd.read_csv("../Data/Raw/NSE_data_all_stocks_2021_upto_31dec2021.csv")
df_2022 = pd.read_csv("../Data/Raw/NSE_data_all_stocks_2022.csv")
df_2023 = pd.read_csv("../Data/Raw/NSE_data_all_stocks_2023.csv")
df_2024 = pd.read_csv("../Data/Raw/NSE_data_all_stocks_2024.csv")


from IPython.display import display

print("2021 Stocks")
display(df_2021.head())
print("2022 Stocks")
display(df_2022.head())
print("2023 Stocks")
display(df_2023.head())
print("2024 Stocks")
display(df_2024.head())



df_2021.rename(columns={
    'DATE': 'Date',
    'CODE': 'Code',
    'NAME': 'Name',
    'Adjust': 'Adjusted Price'
}, inplace=True)

df_2021.head()

## Load Sector Data

import pandas as pd

sectors_2021 = pd.read_csv("../Data/Raw/NSE_data_stock_market_sectors_as_at_31dec2021.csv")
sectors_2022 = pd.read_csv("../Data/Raw/NSE_data_stock_market_sectors_2022.csv")
sectors_2023_2024 = pd.read_csv("../Data/Raw/NSE_data_stock_market_sectors_2023_2024.csv")


print("2021 Sector Data:")
display(sectors_2021.head())
print("\n2022 Sector Data:")
display(sectors_2022.head())
print("\n2023-2024 Sector Data:")
display(sectors_2023_2024.head())

## Standardize Sector Column Names

# Standardize all sector columns
sectors_2021 = sectors_2021.rename(columns={
    'SECTOR': 'Sector', 
    'CODE': 'Code', 
    'NAME': 'Name'
})

sectors_2022 = sectors_2022.rename(columns={
    'Sector': 'Sector',
    'Stock_code': 'Code', 
    'Stock_name': 'Name'
})

sectors_2023_2024 = sectors_2023_2024.rename(columns={
    'Sector': 'Sector',
    'Stock_code': 'Code', 
    'Stock_name': 'Name'
})

## Merge Each Year with Sector Information

# Merge stock data with sectors
df_2021_final = df_2021.merge(sectors_2021[['Code', 'Sector']], on='Code', how='left')
df_2022_final = df_2022.merge(sectors_2022[['Code', 'Sector']], on='Code', how='left')
df_2023_final = df_2023.merge(sectors_2023_2024[['Code', 'Sector']], on='Code', how='left')
df_2024_final = df_2024.merge(sectors_2023_2024[['Code', 'Sector']], on='Code', how='left')

## Concatenate ALL 4 Years

# Concatenate all years
all_stocks = pd.concat([df_2021_final, df_2022_final, df_2023_final, df_2024_final], ignore_index=True)

print(f"\n All 4 years concatenated!")
print(f"  Total rows: {len(all_stocks):,}")
print(f"  Shape: {all_stocks.shape}")
all_stocks.head()

# Checking new structure

print(f"\nMissing values:\n{all_stocks.isna().sum()}")
print(f"\nDuplicates: {all_stocks.duplicated().sum()}")

### Saving merged data

import os
os.makedirs('../data/processed', exist_ok=True)

# Save merged but NOT cleaned data
all_stocks.to_csv('../data/processed/nse_merged_raw.csv', index=False)

print(f"  Rows: {len(all_stocks):,}")
print(f"  Years: 2021-2024")
